+++ board-room/src/main/java/net/stzups/board/server/websocket/MessageHandler.java
@@ -7,14 +7,14 @@ import net.stzups.board.data.objects.Document;
 import net.stzups.board.data.objects.User;
 import net.stzups.board.data.objects.UserSession;
 import net.stzups.board.server.websocket.protocol.client.ClientMessage;
-import net.stzups.board.server.websocket.protocol.client.ClientMessageCreateDocument;
-import net.stzups.board.server.websocket.protocol.client.ClientMessageDraw;
-import net.stzups.board.server.websocket.protocol.client.ClientMessageHandshake;
-import net.stzups.board.server.websocket.protocol.client.ClientMessageOpenDocument;
-import net.stzups.board.server.websocket.protocol.server.ServerMessageAddDocument;
-import net.stzups.board.server.websocket.protocol.server.ServerMessageAddUser;
-import net.stzups.board.server.websocket.protocol.server.ServerMessageDrawClient;
-import net.stzups.board.server.websocket.protocol.server.ServerMessageHandshake;
+import net.stzups.board.server.websocket.protocol.client.messages.ClientMessageCreateDocument;
+import net.stzups.board.server.websocket.protocol.client.messages.ClientMessageDraw;
+import net.stzups.board.server.websocket.protocol.client.messages.ClientMessageHandshake;
+import net.stzups.board.server.websocket.protocol.client.messages.ClientMessageOpenDocument;
+import net.stzups.board.server.websocket.protocol.server.messages.ServerMessageAddDocument;
+import net.stzups.board.server.websocket.protocol.server.messages.ServerMessageAddUser;
+import net.stzups.board.server.websocket.protocol.server.messages.ServerMessageDrawClient;
+import net.stzups.board.server.websocket.protocol.server.messages.ServerMessageHandshake;
 
 import java.util.HashMap;
 import java.util.Map;
