--- board-room/src/main/java/net/stzups/board/server/websocket/protocol/client/ClientMessageCreateDocument.java
@@ -1,7 +1,0 @@
-package net.stzups.board.server.websocket.protocol.client;
-
-public class ClientMessageCreateDocument extends ClientMessage {
-    public ClientMessageCreateDocument() {
-        super(ClientMessageType.CREATE_DOCUMENT);
-    }
-}
