+++ board-room/src/main/java/net/stzups/board/server/websocket/protocol/client/messages/ClientMessageDraw.java
@@ -1,6 +1,8 @@
-package net.stzups.board.server.websocket.protocol.client;
+package net.stzups.board.server.websocket.protocol.client.messages;
 
 import net.stzups.board.data.objects.canvas.Point;
+import net.stzups.board.server.websocket.protocol.client.ClientMessage;
+import net.stzups.board.server.websocket.protocol.client.ClientMessageType;
 
 public class ClientMessageDraw extends ClientMessage {
     private Point[] points;
