+++ board-room/src/main/java/net/stzups/board/server/websocket/protocol/client/messages/ClientMessageHandshake.java
@@ -1,4 +1,7 @@
-package net.stzups.board.server.websocket.protocol.client;
+package net.stzups.board.server.websocket.protocol.client.messages;
+
+import net.stzups.board.server.websocket.protocol.client.ClientMessage;
+import net.stzups.board.server.websocket.protocol.client.ClientMessageType;
 
 public class ClientMessageHandshake extends ClientMessage {
     private long token;
