+++ board-room/src/main/java/net/stzups/board/server/websocket/protocol/server/messages/ServerMessageAddClient.java
@@ -1,7 +1,9 @@
-package net.stzups.board.server.websocket.protocol.server;
+package net.stzups.board.server.websocket.protocol.server.messages;
 
 import io.netty.buffer.ByteBuf;
 import net.stzups.board.server.websocket.Client;
+import net.stzups.board.server.websocket.protocol.server.ServerMessageClient;
+import net.stzups.board.server.websocket.protocol.server.ServerMessageType;
 
 public class ServerMessageAddClient extends ServerMessageClient {
     private Client client;
