+++ board-room/src/main/java/net/stzups/board/server/websocket/protocol/server/messages/ServerMessageAddDocument.java
@@ -1,7 +1,9 @@
-package net.stzups.board.server.websocket.protocol.server;
+package net.stzups.board.server.websocket.protocol.server.messages;
 
 import io.netty.buffer.ByteBuf;
 import net.stzups.board.data.objects.Document;
+import net.stzups.board.server.websocket.protocol.server.ServerMessage;
+import net.stzups.board.server.websocket.protocol.server.ServerMessageType;
 
 public class ServerMessageAddDocument extends ServerMessage {
     private Document document;
