+++ board-room/src/main/java/net/stzups/board/server/websocket/protocol/server/messages/ServerMessageAddUser.java
@@ -1,7 +1,9 @@
-package net.stzups.board.server.websocket.protocol.server;
+package net.stzups.board.server.websocket.protocol.server.messages;
 
 import io.netty.buffer.ByteBuf;
 import net.stzups.board.data.objects.User;
+import net.stzups.board.server.websocket.protocol.server.ServerMessageType;
+import net.stzups.board.server.websocket.protocol.server.ServerMessageUser;
 
 public class ServerMessageAddUser extends ServerMessageUser {
     public ServerMessageAddUser(User user) {
