+++ board-room/src/main/java/net/stzups/board/server/websocket/protocol/server/messages/ServerMessageDrawClient.java
@@ -1,8 +1,10 @@
-package net.stzups.board.server.websocket.protocol.server;
+package net.stzups.board.server.websocket.protocol.server.messages;
 
 import io.netty.buffer.ByteBuf;
 import net.stzups.board.data.objects.canvas.Point;
 import net.stzups.board.server.websocket.Client;
+import net.stzups.board.server.websocket.protocol.server.ServerMessageClient;
+import net.stzups.board.server.websocket.protocol.server.ServerMessageType;
 
 public class ServerMessageDrawClient extends ServerMessageClient {
     private Point[] points;
