+++ board-room/src/main/java/net/stzups/board/server/websocket/protocol/server/messages/ServerMessageOpenDocument.java
@@ -1,9 +1,11 @@
-package net.stzups.board.server.websocket.protocol.server;
+package net.stzups.board.server.websocket.protocol.server.messages;
 
 import io.netty.buffer.ByteBuf;
 import net.stzups.board.data.objects.canvas.Point;
 import net.stzups.board.data.objects.Document;
 import net.stzups.board.data.objects.User;
+import net.stzups.board.server.websocket.protocol.server.ServerMessage;
+import net.stzups.board.server.websocket.protocol.server.ServerMessageType;
 
 import java.util.List;
 import java.util.Map;
