+++ board-room/src/main/java/net/stzups/board/server/websocket/protocol/server/messages/ServerMessageRemoveClient.java
@@ -1,6 +1,8 @@
-package net.stzups.board.server.websocket.protocol.server;
+package net.stzups.board.server.websocket.protocol.server.messages;
 
 import net.stzups.board.server.websocket.Client;
+import net.stzups.board.server.websocket.protocol.server.ServerMessageClient;
+import net.stzups.board.server.websocket.protocol.server.ServerMessageType;
 
 public class ServerMessageRemoveClient extends ServerMessageClient {
     public ServerMessageRemoveClient(Client client) {
